refactor(phase4.1): archive legacy TypeScript type extraction code
Part of the compiler sidecar architecture migration (Phase 4.1).

Archived to _legacy directories:
- ts_check/lib/_legacy/type-extractor.ts
- ts_check/lib/_legacy/type-processor.ts
- ts_check/lib/_legacy/declaration-collector.ts
- ts_check/lib/_legacy/import-handler.ts
- ts_check/lib/_legacy/dependency-manager.ts
- ts_check/lib/_legacy/output-generator.ts
- ts_check/lib/_legacy/type-resolver.ts
- ts_check/_legacy/extract-type-definitions.ts

Added README.md files explaining why code was archived and
how to migrate to the new TypeSidecar architecture.

Updated:
- ts_check/lib/index.ts - reorganized exports, legacy marked deprecated
- ts_check/package.json - added deprecation warning to extract script
- src/analyzer/mod.rs - updated path to archived script location

Kept active (not archived):
- type-checker.ts - still used for type compatibility checking
- manifest-matcher.ts - still used for manifest-based type matching

Type extraction is now handled by TypeSidecar (src/sidecar) using
the compiler-based approach with TypeScript's type inference engine.

All Rust and TypeScript tests pass.
