+++ ts_check/lib/_legacy/type-resolver.ts
