+++ ts_check/lib/index.ts
@@ -1,3 +1,4 @@
+// Active type checking exports (used by sidecar architecture)
 export {
   TypeInfo,
   CompositeAliasInfo,
@@ -7,17 +8,6 @@ export {
 } from "./types";
 export { parseArguments, ParsedArguments } from "./argument-parser";
 export { ProjectManager, isLocalType } from "./project-utils";
-export {
-  findTypeReferenceAtPosition,
-  findTypeDeclarationByPosition,
-  getModuleSpecifierFromNodeModulesPath,
-} from "./type-resolver";
-export { ImportHandler } from "./import-handler";
-export { TypeProcessor } from "./type-processor";
-export { DeclarationCollector } from "./declaration-collector";
-export { DependencyManager } from "./dependency-manager";
-export { OutputGenerator } from "./output-generator";
-export { TypeExtractor } from "./type-extractor";
 export {
   TypeCompatibilityChecker,
   TypeCheckResult,
@@ -65,3 +55,18 @@ export {
   createLogger,
   defaultLogger,
 } from "./logger";
+
+// Legacy exports (archived in _legacy/ directory)
+// These are kept for backward compatibility but are deprecated.
+// Type extraction is now handled by the TypeSidecar (src/sidecar).
+export {
+  findTypeReferenceAtPosition,
+  findTypeDeclarationByPosition,
+  getModuleSpecifierFromNodeModulesPath,
+} from "./_legacy/type-resolver";
+export { ImportHandler } from "./_legacy/import-handler";
+export { TypeProcessor } from "./_legacy/type-processor";
+export { DeclarationCollector } from "./_legacy/declaration-collector";
+export { DependencyManager } from "./_legacy/dependency-manager";
+export { OutputGenerator } from "./_legacy/output-generator";
+export { TypeExtractor } from "./_legacy/type-extractor";
